Renamed files, a lot of editing
Renamed Exercise_1_bash.sh or whatever it was called to BashScript... so
it is easier to find. Also made a ton of edits on PyScript where it is
properly translating the sequences, and counting the variant sites. Now
I just need to create a new fasta file with the protein sequences.